Added experimental module readers for plain and s-expression Javascript.
Each language creates a module that defines and exports a single variable, "ans", containing a Javascript !begin block:

  #lang planet untyped/mirrors/javascript/plain
  function sum (a, b) { return a + b; }

  #lang planet untyped/mirrors/javascript/sexp
  (function sum (a b) (return (+ a b)))

The idea is to create a require-like mechanism that allows the programmer to specify dependencies between modules.
Once we have that, we have the basis for an automated JS build process.

And if it works with JS, it can work with CSS too.