Added a very simple Javascript module system.
Any JS file written in the s-exp language (not the plain
language, yet) exports a single binding called 'script',
which contains the code as a Javascript structure. This
structure is also placed in Javascript registry
(javascript-registry.ss) where it can be later recalled
using registry->string. As modules are always loded in
topological order they will always be registered in the
correct order so dependencies come before scripts that
depend on them.

I have one final change to make: require statements for
Javascript inside a Javascript file should automatically
prefix their import so it doesn't clash with the script
binding created in the requiring module.

Note that javascript-registry.ss is written in Typed Scheme.